+++ metric/internal/global/instruments.go
@@ -18,8 +18,8 @@ import (
 	"context"
 	"sync/atomic"
 
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
+	oGlob "go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/instrument"
 )
@@ -44,7 +44,7 @@ var _ instrument.Float64ObservableCounter = (*afCounter)(nil)
 func (i *afCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -72,7 +72,7 @@ var _ instrument.Float64ObservableUpDownCounter = (*afUpDownCounter)(nil)
 func (i *afUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -100,7 +100,7 @@ var _ instrument.Float64ObservableGauge = (*afGauge)(nil)
 func (i *afGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -128,7 +128,7 @@ var _ instrument.Int64ObservableCounter = (*aiCounter)(nil)
 func (i *aiCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -156,7 +156,7 @@ var _ instrument.Int64ObservableUpDownCounter = (*aiUpDownCounter)(nil)
 func (i *aiUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -184,7 +184,7 @@ var _ instrument.Int64ObservableGauge = (*aiGauge)(nil)
 func (i *aiGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -212,7 +212,7 @@ var _ instrument.Float64Counter = (*sfCounter)(nil)
 func (i *sfCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Counter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -238,7 +238,7 @@ var _ instrument.Float64UpDownCounter = (*sfUpDownCounter)(nil)
 func (i *sfUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -264,7 +264,7 @@ var _ instrument.Float64Histogram = (*sfHistogram)(nil)
 func (i *sfHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Histogram(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -290,7 +290,7 @@ var _ instrument.Int64Counter = (*siCounter)(nil)
 func (i *siCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Counter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -316,7 +316,7 @@ var _ instrument.Int64UpDownCounter = (*siUpDownCounter)(nil)
 func (i *siUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -342,7 +342,7 @@ var _ instrument.Int64Histogram = (*siHistogram)(nil)
 func (i *siHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Histogram(i.name, i.opts...)
 	if err != nil {
-		otel.Handle(err)
+		oGlob.GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
